Delete generation of Pyglet window
The program was opening an additional "main.py" window (blank, just a black square) and if I clicked/tapped in that square it affected how the alarm is playing back (it seemed to switch to playing the full audio file instead of looping the first portion). I was thinking it would be good to suppress that window if possible... and took a guess that commenting out this line might work and it seemed to do the trick - no more black window but alarms still seem to work!