sub 1 inH2O = 0
Changed the conversion of theta to inH2O to treat values under 1 inH2O as 0 inH2O. Again, this is an artifact of the when the needle is at its lowest point, it is pretty close to the 1 inH2O mark, such that it visually reads as somewhere between .5 and .9.  Rather than just passing on this value, I think it's better to say: IF the inH2O value is less than 1, it treats it as 0; otherwise it returns the real inH2O value -- then we advise the user that it doesn't really measure pressure in the sub 1 inH2O (~ sub 2.5 cmH2O) range.